# src/main/resources/swagger/openapi-doc.yaml
openapi: 3.0.1
info:
  title: Bazario marketplace Panel API
  description: |
    The Bazario marketplace Panel API is a comprehensive RESTful API
    that serves as the backbone of the marketplace`s digital platform.
    It provides a wide range of endpoints for managing crud-operation of ad, users
    and admins.

  version: v0.0.1
servers:
  - url: https://bazario-mkur.onrender.com
    description: Remote service for managing platform of the Bazario marketplace
  - url: http://127.0.0.1:8080
    description: Local service for managing platform of the Bazario marketplace
...
paths:
...
  /public/refreshToken:
    post:
      tags:
        - Authentication
      description: User login endpoint
      operationId: refreshToken
      parameters:
        - in: cookie
          name: refreshToken
          schema:
            type: string
      responses:
        200:
          description: Authenticated successfully. Get token.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JWT_REFRESH_TOKEN=abcs123
          content:
            text/plain:
              schema:
                type: string
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        429:
          description: Too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      description: Error message with timestamp
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT